+++ sdk/dir-js/src/models/sign_v1.ts
